Now the spacing is regular!!
It's not correct yet, but this is an amazing step. We were making some kind of error that I did not directly fix. Instead, I tried rewriting the existing code by using two separate sets of output variables for the massive and massless results. Now the code is more readable and the error evaporated!