New approach to digesting models
Instead of hand-copying everything, we now simply use Pandas and treat the cosmology .dat file as a whitespace-separated CSV.

Unfortunately, this does not seem to have affected the results in any way. However, it was responsible to do this in any case, because we have now eliminated the possibility that our woes were due to some obscure copy error.