Upload Cake 22 notebooks
I really should have done this earlier, maybe then I would have had some presentable results for the meeting today.

The desktop has 16 cores, and its performance is less important to me anyway. Why don't we nuke it with ML? But we should be careful to save our results this time (or rather, incrementally create new notebooks with what we learn), so that we could present on what we learn.