Fix redundant pathing
The generate_training_data script was reading in the cosmologies data table again, which is simply adding a point of failure in case we decide to path the file differently. Now it uses the same cosm object as in ci, which is more efficient because ci is already imported anyway!