Fix cosmologies.dat pathing once and for all
Now, when the user installs the cassL package, a copy of cosmologies.dat is automatically written to the installation location. To reference this file, camb_interface now looks correctly within the same directory.

With this fix, we will never have to modify the scripts solely to load the cosmologies table. Furthermore, even if there were a problem for some reason, the code should theoretically still be able to run without it!

To toast this development, we're upgrading to version 0.0.2