Clear out debug and Andrea code
We're clearing these things at least out of the interpolate cell function. The function itself doesn't require that much intense debugging anymore.

Eventually, I hope to do the same with the evaluate_cell function. But I would prefer to train and test an emulator using that function before I remove the debug code, because that approach is not yet guaranteed to be fully correct from start to finish--as of now, we've only verified its accuracy in the new_CAMB_agreement sandbox.

Also in this commit: we've fixed a formatting bug which caused the script to crash upon encountering hopeless cells. Before, we were returning None for the rescaling parameters when it was more appropriate to return an array of Nones. The fix uses an array of NaNs because an array of Nones would be an object array, which I want to avoid as much as I can.