Small expansions in the outline
We've also introduced some fairly useful code in the quick run script for picking up from where we left off. I hope we can include that in the other script eventually, too.

Also in this commit: a clear roadmap, all of the emulators we need to collect on a flashdrive or something so that we're not spinning our wheels when we switch platforms.