Several New Sections
Partially we've transcribed the Otto notes from yesterday. Partially we've transcribed the notes from the phone.

Several sections have been split up, several sections represent experiments with our emulator that we have yet to run.

We've expanded the number of experiments described in complaints.txt, and revised some of them to be more realistic. We have many now, and we hope to have all experiments simultaneously accessible on all devices. Therefore, we should focus the coming commits on the train_emu and user_interface scripts.

One idea for an expansion: resume_building function will automatically look for the last backup file associated with a label, then will build it unto completion. It should auto-detect properties like the eventual capacity of the hypercube.

Otherwise, these scripts will simply be expanded by condensing the quick_run scripts as well as the code included in the emulator notebooks!