Fix insensitivity to num_samples
I thought I was computing the larger hypercubes, but they were all actually of size 5k, because the script was not correctly using the parameter.

Also in this commit: emulator experiment roadmap is now much more detailed and structured in the form of a to-do list.

Also: section titles are clearer in the outline