Protect file handling in quick run
Now the completed_index variable is -1 instead of None, meaning it is more versatile. Furthermore, if completed_index is not -1 but we don't have a relevant backup, we reset it to -1.

This quick-run script has become quite involved, so  we now have even greater pressure to develop this train_emu script.