Finish test function, get started on error_plot
The first draft of the error_plot function is not yet ready, we still need to implement the fixed_k functionality. Then we'll be able to to move on to the  user_interface functions.

Also in this commit: added some dud code to keep the compiler from complaining about the user_interface script when we import cassL.