Fix several unclosed error-raises
Also in this commit: we've added the infamous try / catch back into the generate_emu_data script. We needed to do this because the MEGA priors caused a CAMBError crash, and we lost about 430 samples.

I didn't actually test to make sure that this really catches the CAMBError, we may have to take the script apart a little more carefully