Generalize dimension checks
Previously, we were referring to the shapes of the X and Y training sets in order to perform some input checking. It's better if we include xdim and ydim attributes in the emulator object, because some emulators will be loaded directly from a file and will actually lack these data sets!

Next up:

1. Continue to test build_and_test_emulator function on some of the hypercubes we've been making so far

2. Complete the build_train_and_test_sets function

3. Add an even higher-level function called build_emulator_from_scenario which unites build_and_test_emulator with build_train_and_test_sets.