Add support for different standard_k sizes
This will allow us to add a section in the thesis about varying the number of points in k.

We just need one more set of code modifications in order to get the LHC tests available. I'll do that later, though, because the computers are already busy pushing out results.