Extreme introduction expansions
We have a new section (mamma mia!) serving as a glossary for the various cosmological parameters.

In particular, we're in the middle of expanding the following three sections:
1. Zusammenfassung
2. Glossary of parameters
3. Power spectrum

It's looking good so far. The manuscript is a mess right now and won't compile, but we haven't committed in a while, so we'll patch things up in the next commit.