Intro + Zusammenfassung cleanup
The last commit was incorrect, actually the document did compile. It just encountered several compiler errors.

The compiler errors have been fixed and we've added a few more expansions.

For now, we should pause and finish incorporating content from the Google Doc. It's getting increasingly difficult to follow this document on phone.