Major reorganization of intro
We've written some new material for the glossary of parameters. Most importantly, all of the currently-existing introduction material is filed under the correct section. The intro is still a mess, but it is a mess with a kind of logical flow.

Also in this commit: several missing section labels were fixed.