Constrain scope of ci coverage
We're only mentioning some of the functions from camb_interface (ci) in this thesis. Many functions are outdated, some were written by Andrea, and at least one served only to debug. All of the functions that I've written anywhere in the document should cover everything.

We've made a little progress on 2.1 but it's still rough. I think we'll be able to pull it together quickly tomorrow...