New table describing cosmology dict
There's a problem with it, though--the footnote isn't working. We should look this up.

Also in this commit: a default_cosmology function, which should make camb_interface more user-friendly because we aren't relying on the user to protect Aletheia model 0.

We've still got a good amount of writing to do for chapter 2, but I think we've made it past all discussion of CAMB's peculiarities...