Flow chart
This was by far the most difficult part of chapter 4. The rest of the material should flow relatively easily from this...

For the next commit, we should concentrate on incorporating my written feedback on the sections 4.2-4.4. Then we should go back and start writing 4.1 as a breakdown of the flow chart.

Try not to talk about what each script does as much as what flow chart deliverable we want from it.