Expand ch4 intro, switch 4.1 and 4.2.
When I started to write 4.1 (which at this time was the flow chart section), I realized I kept referring to LHSs and prior files. I am therefore switching 4.1 with 4.2, even though I find it highly ugly to place the flow chart section in the middle of the chapter. Maybe Ariel will have an idea about how to do this better, so I'm not going to worry too much about the ugliness right now