Revise chapter 5.6
We need to write a brief introduction for the chapter, including installation instructions.

Also, chapter 5.6 ends a little abruptly. Maybe we can come back and segue more fluidly into chapter 6?

Then we can move on to chapter 6. But that means that we've reached chapter 6 before the computers finished generating the results! Yikes.

Well, we'll think of something. The computers are really close to being done... maybe we can kill time by writing the summary of the thesis or cleaning up the future work section.