Move 1.8 into chapter 4
This LHS section is HUGE, but I probably won't split it up into separate LHS sections.

We need to integrate 1.8 better, right now it's mostly just sitting there.

We've started to write the conclusion for the paper, but most of it is just a duplicate of the abstract. The conclusion and future work sections are the most promising at the moment because they only depend on the not-yet-available plots in very tight regions.