Clear out chapter 5
Remember that chapter 6 ultimately should not have any sections. The sections that are currently in the text only serve to distinguish to me the different threads of conversation. Ultimately, everything must go in either the summary or the future work.

We should ask Ariel for input on the future work section. I have a feeling it will go quite long, so I will need his input about whether I should toss anything out for being inconsequential!