Remove symbolic regression from future work
I've decided it makes more sense if I finish that thought when it first appears in chapter 3: it's not a good idea, and I want to put it out of the reader's mind.

Now we will continue to clean up and tighten the future work section. The thesis is really almost done!