Caption for figure 4.3
Also in this commit: move the code for the flow chart so that the other chapter-four figures don't get kicked.

After the "---", the thesis would be acceptable for submission, if you so desired.

Next up:
* Redo plot for chapter 2
* Rewrite chapter 2 with material from 1.7 on the use
of tilde. You'll have to include new plots just for that! as well as a break out a new section 2.1.
* Create command to convert orange text to comment. Convert orange text to commands
* Also check for ?? and other TeX warnings
* Sweep the thesis for missing captions.
* Rewrite LHS material so that it is a cohesive section
---
* Small revisions
* Correct the "Medians Absolute Error" typo and add $\sigma_{12}$ to the color bar label.
* If there's extra times: regenerate the plots that Ariel asked for