Rewrite LHS sections
In other words, we finally redid the plot in chapter 2. I wonder why the work desktop cannot solve this
notebook correctly...

After the "---", the thesis would be acceptable for submission, if you so desired.

Next up:
* Rewrite chapter 2 with material from 1.7 on the use
of tilde. You'll have to include new plots just for that! as well as a break out a new section 2.1.
* Create command to convert orange text to comment. Convert orange text to commands
* Also check for ?? and other TeX warnings
* Sweep the thesis for missing captions.
---
* Small revisions
* Correct the "Medians Absolute Error" typo and add $\sigma_{12}$ to the color bar label.
* If there's extra times: regenerate the plots that Ariel asked for