Sections 2.1 and 2.2 revised
There's a bit more cleanup to do in chapter 2, for example the summary needs to mention section 2.1. Also, a couple of caption complaints are out of date. Then we can move on to other to-do items!

After the "---", the thesis would be acceptable for submission, if you so desired.

Next up:
>> Finish section 2.1 and rewrite 2.2 intro to flow.
* Create command to convert orange text to comment. Convert orange text to commands
>>> At this point get ready for bed.
* Also check for ?? and other TeX warnings
* Sweep the thesis for missing captions.
* Use some include statements to get the required paperwork into the thesis.
---
* Small revisions
* Correct the "Medians Absolute Error" typo and add $\sigma_{12}$ to the color bar label.
* If there's extra times: regenerate the plots that Ariel asked for