Chapter 3 swept for color text
I've decided not to check for captions after all. There may be some embarrassing blunders left in the thesis, but Marion is far more important to me than this.

After the "---", the thesis should simply be submitted. But the remaining points are important for creating a version with which I'm happy.

Next up:
Sweep chapters for color text
>> 4
* 5
* 6
### At this point, get ready for bed.
* Use some include statements to get the required paperwork into the thesis.
---
* Sweep the thesis for missing captions.
* Correct the "Medians Absolute Error" typo and add $\sigma_{12}$ to the color bar label.
* Remaining revisions Ariel wrote on the print-outs.
Especially:
* If there's extra times: regenerate the plots that Ariel asked for, such as the six-panel deluxe in chapter 1.