Fix axis bug in denormalize_row
Now the xrange and xmin should be calculated correctly.

Also in this commit: we're in the middle of updating the train_emu script for training over uncertainties. Unfortunately, this looks like it will be more work than we expected: the Emulator class already assumes Y will have to be denormalized, which doesn't work if we're training over uncertainties (I don't think normalization will be necessary there).