Priors re-tightened
We hit yet another failed cosmology, so we're tightening again. Conservatism increased from 75% to 96%!! ;(

At this point, the prior-widening experiment can safely be considered a failure. We still have over 6700 cosmologies to analyze, so there's a very good chance that the priors could be tightened even further.

Even if we get lucky and make it through these remaining cosmologies without another adjustment, a 96% conservatism level means that we've failed to widen the priors by any significant margin.

Nevertheless, as long as the final conservatism is less than 100%, I'll still replace the old emulator with a new one.