Added ultra-miserly h modification clauses
Before, the generate_emu_data routines would call it quits if the cosmology could not be solved with h=0.07. However, in practice, I found that sometimes pushing h even smaller could still work... it would simply enter unstable territory. For reference, CAMB sets a hard limit at h=0.01--anything smaller and CAMB won't even attempt to solve it. By contrast, if, say, h=0.01, CAMB will attempt to solve it, and you'll get some other error which is still due to the tiny h value but which does not explicitly complain about h.

Now, the code will push CAMB to the limits (at least, at the percent level), and squeeze all of the solving power out of the h-variation procedure that we can get.

Also in this commit:
* Break a new fn broadcast_unsolvable out of direct_eval_cell and interpolate_cell. This code handles the reporting of unsolvable cells.
* quickrun_h_haggle allows us to quickly test the new h-variation code.