Adjustments to conservatism investigation
We zoomed in on index 3248 of the validation set: the most challenging cosmology that we've yet encountered. This allows us to very quickly determine the new conservatism level.

The ultra-miserly h-variation code proved surprisingly brilliant. By allowing h to vary now also from 0.07 to 0.02, we've managed to reduce conservatism from 96% to 39%! A wild success.

Also in this commit:
* New quickrun script: quickrun_prior_widener. It's basically quick_run_Hnu with fewer options, in order to maintain simplicity.