Critical prior_tightener bug fixed
We were re-using the hc label even though that cosmology's sigma12 field was getting mutated by the fill_hypercube calls!!

With this fix, the results of prior_tightener remain valid when plugged into quickrun_h_haggle.

Two commits ago, we reported that the new conservatism is 39%. However, this number was erroneously low because of the prior_tightener bug addressed in this commit. We've re-run the script and determined that the correct conservatism is 45%!