Workaround for ultra-rare bug
We've encountered an extremely bizarre situation where SciPy will throw a ValueError (complaining about real sigma12 being outside the interpolator's range) so powerful, that my "except ValueError:" statement is unable to actually catch it, and it slips a few statements further into the code.

This happened for index 593 of the validation set. It reached h=0.05 and then obtained a value of sigma12 that was outside of the interpolator range by an amount due seemingly exclusively to floating point error--the complaint was something like value 1.000000001 being above the maximum interpolator value of 1.0.

Even if this is an admittedly rare borderline case, the mere fact that it's borderline does nothing to explain why SciPy's ValueError becomes magical. Keep in mind that the code was literally designed to catch this specific ValueError (the sigma12 being outside the interpolator range), because it tells the evaluation function that we need to decrease h (or, declare an unsolvable cell if we can't decrease h any further--but in this case, we certainly could have decreased h further to solve the problem, and indeed I tested that the h=0.04 case is handled just fine).

My solution to this problem is INCREDIBLY ugly and hacky. I'm simply accepting that the ValueError is magical and slips further / out of the recursion. I've added an additional except statement at the point to which it's consistently slipping. I've copied some of the outer code so that If I catch the error here, I simply decrease h and try again. The hope is that decreasing h will dislodge us from this horrible borderline behavior.