Testing bug resolved
This has happened a surprising number of times in the development of this project. I want to check some function's output to make sure it's correct--I test it separately from the main script. It looks wrong because the testing script itself has the bug. But because the testing script has at least two bugs, it's not obvious that the testing script is at fault.

In this case, the two bugs were:
* I wasn't testing like cosmologies at the same time, I was accidentally testing Aletheia model 0 against the first row of the training lhc.
* The small_k array didn't span the same range as the big_k array. The small_k array started at 1e-4 and ended at 10, the big_k started at 7e-5 and ended at 7. This led to a shifting that miraculously looked almost exactly like a factor of h shift! That's because 7 / 10 is almost the same as 0.67.

What a terrible coincidence...

Anyway, the investigation has come to a conclusion. All four founctions agree on at least one cosmology, so I'm going back to computing masses of spectra!!