First draft of full pipeline code
The giant LHC is still in the process of generating, so we're going to switch to the MPE desktop and test this code on a simpler case of 1000 cosmologies. Even though this will cover very little of the ten-dimensional input space, it should still tell us if the code is incorrect.