Fix bad indexing and missing names
We've also added a couple empty lines in the output, which is definitely unclean because it assumes that crunch_and_sew will print in a particular way. After all, I may decide to revise that later...

Anyway, the code is now at least running. We'll have to wait and see if the outputs are any good, though.