Collapse get_CAMB_sigma12 and evaluate_sigmaR
It was too confusing to have two functions with basically the same name.

Also in this commit: first draft of ci's get_sigmaR fn. I'm committing now so I can play around a bit and understand its outputs better. If everything checks out, we can proceed to writing a ged fn to call this for the purposes of building a sigmaR emulator!