Split fill_hypercube into two fn.s
Although the functions share a good amount of code, squishing them together makes the function too cumbersome and illegible, I feel. This issue came to a head because I was trying to squeeze in yet a third case.

In the next commit, I'll implement this third case, this time as a separate fn in the vein of these two here.

Also in this commit: documentation fix for get_sigmaR