Re-do the 2-percent priors
It would be more useful for the audience if the 2% scenario led to errors for the entire pipeline, not just the emulator.

This means that we'll have to separately enforce the sigma12 constraint later, though.... Furthermore, if the DE situation never gets sorted out, these errors might still be pretty bad on large scales. Let's take things one step at a time.