Break out new visualizations script
It may need to be renamed to something like "demonstrations," but anyway it's helpful for segregating code that is extremely low priority for maintenance. This kind of code is sort of one-purpose, but I'll try to write it so that it can still be called later.

Also in this commit:
* Discard a couple of functions that have long since vanished from our vocabulary.
* Rewrite some documentation because we have a better understanding now of the field and we don't need to ask so many questions.